refactor: rename and reorganize color correction module structure
