+++ .gitignore
@@ -20,4 +20,4 @@ wheels/
 
 tmp/
 zzz/
-color_correction_asdfghjkl/asset/images/
+color_correction/asset/images/
