+++ Makefile
@@ -1,6 +1,6 @@
 yolo-export-onnx:
 	yolo export \
-	model=color_correction_asdfghjkl/asset/.model/yv8-det.pt \
+	model=color_correction/asset/.model/yv8-det.pt \
 	format=onnx \
 	device=mps \
 	simplify=True \
