+++ color_correction/constant/yolov8_det.py
