+++ color_correction/core/correction/__init__.py
@@ -1,0 +1,17 @@
+from color_correction.core.correction._factory import (
+    CorrectionModelFactory,
+)
+from color_correction.core.correction.affine_reg import AffineReg
+from color_correction.core.correction.least_squares import (
+    LeastSquaresRegression,
+)
+from color_correction.core.correction.linear_reg import LinearReg
+from color_correction.core.correction.polynomial import Polynomial
+
+__all__ = [
+    "CorrectionModelFactory",
+    "LeastSquaresRegression",
+    "Polynomial",
+    "LinearReg",
+    "AffineReg",
+]
