+++ color_correction/core/correction/_factory.py
@@ -1,9 +1,9 @@
-from color_correction_asdfghjkl.core.correction.affine_reg import AffineReg
-from color_correction_asdfghjkl.core.correction.least_squares import (
+from color_correction.core.correction.affine_reg import AffineReg
+from color_correction.core.correction.least_squares import (
     LeastSquaresRegression,
 )
-from color_correction_asdfghjkl.core.correction.linear_reg import LinearReg
-from color_correction_asdfghjkl.core.correction.polynomial import Polynomial
+from color_correction.core.correction.linear_reg import LinearReg
+from color_correction.core.correction.polynomial import Polynomial
 
 
 class CorrectionModelFactory:
