+++ color_correction/core/correction/least_squares.py
@@ -1,35 +1,30 @@
 import time
 
 import numpy as np
-from sklearn.linear_model import LinearRegression
-from sklearn.pipeline import make_pipeline
-from sklearn.preprocessing import PolynomialFeatures
 
-from color_correction_asdfghjkl.core.correction.base import BaseComputeCorrection
-from color_correction_asdfghjkl.utils.correction import (
+from color_correction.core.correction.base import BaseComputeCorrection
+from color_correction.utils.correction import (
     postprocessing_compute,
     preprocessing_compute,
 )
 
 
-class Polynomial(BaseComputeCorrection):
-    def __init__(self, **kwargs: dict) -> None:
+class LeastSquaresRegression(BaseComputeCorrection):
+    def __init__(self) -> None:
         self.model = None
-        self.degree = kwargs.get("degree", 2)
 
     def fit(
         self,
         x_patches: np.ndarray,  # input patches
         y_patches: np.ndarray,  # reference patches
-        **kwargs: dict,
     ) -> np.ndarray:
         start_time = time.perf_counter()
 
-        degree = kwargs.get("degree", self.degree)
-        self.model = make_pipeline(
-            PolynomialFeatures(degree),
-            LinearRegression(),
-        ).fit(x_patches, y_patches)
+        self.model = np.linalg.lstsq(
+            a=x_patches,
+            b=y_patches,
+            rcond=None,
+        )[0]  # get only matrix of coefficients
 
         exc_time = time.perf_counter() - start_time
         print(f"{self.__class__.__name__} Fit: {exc_time} seconds")
@@ -39,8 +34,9 @@ class Polynomial(BaseComputeCorrection):
         if self.model is None:
             raise ValueError("Model is not fitted yet. Please call fit() method first.")
 
+        # Input adalah array (N,3) dari nilai warna patches
         org_input_shape = input_image.shape
         input_image = preprocessing_compute(input_image)
-        image = self.model.predict(input_image)
+        image = np.dot(input_image, self.model)
         corrected_image = postprocessing_compute(org_input_shape, image)
         return corrected_image
