+++ color_correction/core/correction/polynomial.py
@@ -1,30 +1,35 @@
 import time
 
 import numpy as np
+from sklearn.linear_model import LinearRegression
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import PolynomialFeatures
 
-from color_correction_asdfghjkl.core.correction.base import BaseComputeCorrection
-from color_correction_asdfghjkl.utils.correction import (
+from color_correction.core.correction.base import BaseComputeCorrection
+from color_correction.utils.correction import (
     postprocessing_compute,
     preprocessing_compute,
 )
 
 
-class LeastSquaresRegression(BaseComputeCorrection):
-    def __init__(self) -> None:
+class Polynomial(BaseComputeCorrection):
+    def __init__(self, **kwargs: dict) -> None:
         self.model = None
+        self.degree = kwargs.get("degree", 2)
 
     def fit(
         self,
         x_patches: np.ndarray,  # input patches
         y_patches: np.ndarray,  # reference patches
+        **kwargs: dict,
     ) -> np.ndarray:
         start_time = time.perf_counter()
 
-        self.model = np.linalg.lstsq(
-            a=x_patches,
-            b=y_patches,
-            rcond=None,
-        )[0]  # get only matrix of coefficients
+        degree = kwargs.get("degree", self.degree)
+        self.model = make_pipeline(
+            PolynomialFeatures(degree),
+            LinearRegression(),
+        ).fit(x_patches, y_patches)
 
         exc_time = time.perf_counter() - start_time
         print(f"{self.__class__.__name__} Fit: {exc_time} seconds")
@@ -34,9 +39,8 @@ class LeastSquaresRegression(BaseComputeCorrection):
         if self.model is None:
             raise ValueError("Model is not fitted yet. Please call fit() method first.")
 
-        # Input adalah array (N,3) dari nilai warna patches
         org_input_shape = input_image.shape
         input_image = preprocessing_compute(input_image)
-        image = np.dot(input_image, self.model)
+        image = self.model.predict(input_image)
         corrected_image = postprocessing_compute(org_input_shape, image)
         return corrected_image
