+++ color_correction/utils/image_patch.py
