+++ color_correction/utils/image_processing.py
