+++ tests/core/card_detection/test_yolov8_detector.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from color_correction_asdfghjkl.core.card_detection.det_yv8_onnx import YOLOv8CardDetector
+from color_correction.core.card_detection.det_yv8_onnx import YOLOv8CardDetector
 
 @pytest.mark.skip(reason="Test is not implemented")
 def test_detector_init(sample_image: np.ndarray) -> None:
